@@ -1,11 +1,11 @@
 name: notus
 description: Platform-agnostic rich text document model based on Delta format and used in Zefyr editor.
-version: 0.1.4
+version: 0.1.5
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 homepage: https://github.com/memspace/zefyr
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.2.0 <3.0.0'
 
 dependencies:
   collection: ^1.14.6
@@ -15,5 +15,5 @@ dependencies:
 
 dev_dependencies:
   pedantic: ^1.0.0
-  test: ^1.0.0
-  test_coverage: ^0.2.1
+  test: ^1.10.0
+  test_coverage: ^0.4.0
