@@ -38,6 +38,7 @@ abstract class NotusAttributeKey<T> {
 abstract class NotusAttributeBuilder<T> implements NotusAttributeKey<T> {
   const NotusAttributeBuilder._(this.key, this.scope);
 
+  @override
   final String key;
   final NotusAttributeScope scope;
 
@@ -139,9 +140,11 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   const NotusAttribute._(this.key, this.scope, this.value);
 
   /// Unique key of this attribute.
+  @override
   final String key;
 
   /// Scope of this attribute.
+  @override
   final NotusAttributeScope scope;
 
   /// Value of this attribute.
@@ -160,6 +163,7 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   ///
   /// When composed into a rich text document, unset attributes remove
   /// associated style.
+  @override
   NotusAttribute<T> get unset => NotusAttribute<T>._(key, scope, null);
 
   /// Returns `true` if this attribute is an unset attribute.
@@ -168,6 +172,7 @@ class NotusAttribute<T> implements NotusAttributeBuilder<T> {
   /// Returns `true` if this is an inline-scoped attribute.
   bool get isInline => scope == NotusAttributeScope.inline;
 
+  @override
   NotusAttribute<T> withValue(T value) =>
       NotusAttribute<T>._(key, scope, value);
 
@@ -206,7 +211,7 @@ class NotusStyle {
     return NotusStyle._(result);
   }
 
-  NotusStyle() : _data = Map<String, NotusAttribute>();
+  NotusStyle() : _data = <String, NotusAttribute>{};
 
   /// Returns `true` if this attribute set is empty.
   bool get isEmpty => _data.isEmpty;
@@ -402,6 +407,7 @@ class EmbedAttributeBuilder
   @override
   NotusAttribute<Map<String, dynamic>> get unset => EmbedAttribute._(null);
 
+  @override
   NotusAttribute<Map<String, dynamic>> withValue(Map<String, dynamic> value) =>
       EmbedAttribute._(value);
 }
