@@ -2,26 +2,43 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
 import 'src/form.dart';
 import 'src/full_page.dart';
 import 'src/view.dart';
+import 'src/text_field_page.dart';
 
 void main() {
   runApp(ZefyrApp());
+}
+
+// Create a Focus Intent that does nothing
+class FakeFocusIntent extends Intent {
+  const FakeFocusIntent();
 }
 
 class ZefyrApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      // disable default arrows focus changes.
+      // otherwise it makes the keyboard flicker when we move with arrows)
+      shortcuts: Map<LogicalKeySet, Intent>.from(WidgetsApp.defaultShortcuts)
+        ..addAll(<LogicalKeySet, Intent>{
+          LogicalKeySet(LogicalKeyboardKey.arrowLeft): const FakeFocusIntent(),
+          LogicalKeySet(LogicalKeyboardKey.arrowRight): const FakeFocusIntent(),
+          LogicalKeySet(LogicalKeyboardKey.arrowDown): const FakeFocusIntent(),
+          LogicalKeySet(LogicalKeyboardKey.arrowUp): const FakeFocusIntent(),
+        }),
       debugShowCheckedModeBanner: false,
       title: 'Zefyr Editor',
-      theme: ThemeData(primarySwatch: Colors.cyan),
       home: HomePage(),
       routes: {
         "/fullPage": buildFullPage,
         "/form": buildFormPage,
         "/view": buildViewPage,
+        "/textinput": buildTextFieldPage,
       },
     );
   }
@@ -37,6 +54,10 @@ class ZefyrApp extends StatelessWidget {
   Widget buildViewPage(BuildContext context) {
     return ViewScreen();
   }
+
+  Widget buildTextFieldPage(BuildContext context) {
+    return TextFieldScreen();
+  }
 }
 
 class HomePage extends StatelessWidget {
@@ -44,32 +65,25 @@ class HomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     final nav = Navigator.of(context);
     return Scaffold(
-      appBar: AppBar(
-        elevation: 1.0,
-        backgroundColor: Colors.grey.shade200,
-        brightness: Brightness.light,
-        title: ZefyrLogo(),
-      ),
+      appBar: AppBar(title: ZefyrLogo()),
       body: Column(
         children: <Widget>[
           Expanded(child: Container()),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/fullPage'),
             child: Text('Full page editor'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
           ),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/form'),
             child: Text('Embedded in a form'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
           ),
-          FlatButton(
+          RaisedButton(
             onPressed: () => nav.pushNamed('/view'),
             child: Text('Read-only embeddable view'),
-            color: Colors.lightBlue,
-            textColor: Colors.white,
+          ),
+          RaisedButton(
+            onPressed: () => nav.pushNamed('/textinput'),
+            child: Text('basic text input'),
           ),
           Expanded(child: Container()),
         ],
