@@ -93,7 +93,7 @@ class ZefyrButton extends StatelessWidget {
     }
   }
 
-  Color _getColor(ZefyrScope editor, ZefyrToolbarTheme theme) {
+  Color _getColor(ZefyrScope editor, ToolbarTheme theme) {
     if (isAttributeAction) {
       final attribute = kZefyrToolbarAttributeActions[action];
       final isToggled = (attribute is NotusAttribute)
@@ -306,7 +306,6 @@ class _LinkButtonState extends State<LinkButton> {
   final TextEditingController _inputController = TextEditingController();
   Key _inputKey;
   bool _formatError = false;
-  ZefyrScope _editor;
 
   bool get isEditing => _inputKey != null;
 
@@ -432,7 +431,7 @@ class _LinkButtonState extends State<LinkButton> {
     final toolbar = ZefyrToolbar.of(context);
     final style = toolbar.editor.selectionStyle;
 
-    String value = 'Tap to edit link';
+    var value = 'Tap to edit link';
     if (style.contains(NotusAttribute.link)) {
       value = style.value(NotusAttribute.link);
     }
@@ -528,7 +527,7 @@ class _LinkInputState extends State<_LinkInput> {
     final toolbarTheme = ZefyrTheme.of(context).toolbarTheme;
     final color =
         widget.formatError ? Colors.redAccent : toolbarTheme.iconColor;
-    final style = theme.textTheme.subhead.copyWith(color: color);
+    final style = theme.textTheme.subtitle1.copyWith(color: color);
     return TextField(
       style: style,
       keyboardType: TextInputType.url,
@@ -568,7 +567,7 @@ class _LinkView extends StatelessWidget {
               value,
               maxLines: 1,
               overflow: TextOverflow.ellipsis,
-              style: theme.textTheme.subhead
+              style: theme.textTheme.subtitle1
                   .copyWith(color: toolbarTheme.disabledIconColor),
             ),
           )
