@@ -21,6 +21,13 @@ class ZefyrField extends StatefulWidget {
   final ZefyrImageDelegate imageDelegate;
   final ScrollPhysics physics;
 
+  /// The appearance of the keyboard.
+  ///
+  /// This setting is only honored on iOS devices.
+  ///
+  /// If unset, defaults to the brightness of [ThemeData.primaryColorBrightness].
+  final Brightness keyboardAppearance;
+
   const ZefyrField({
     Key key,
     this.decoration,
@@ -32,6 +39,7 @@ class ZefyrField extends StatefulWidget {
     this.toolbarDelegate,
     this.imageDelegate,
     this.physics,
+    this.keyboardAppearance,
   }) : super(key: key);
 
   @override
@@ -39,6 +47,7 @@ class ZefyrField extends StatefulWidget {
 }
 
 class _ZefyrFieldState extends State<ZefyrField> {
+  ZefyrMode get _effectiveMode => widget.mode ?? ZefyrMode.edit;
   @override
   Widget build(BuildContext context) {
     Widget child = ZefyrEditor(
@@ -46,10 +55,11 @@ class _ZefyrFieldState extends State<ZefyrField> {
       controller: widget.controller,
       focusNode: widget.focusNode,
       autofocus: widget.autofocus,
-      mode: widget.mode ?? ZefyrMode.edit,
+      mode: _effectiveMode,
       toolbarDelegate: widget.toolbarDelegate,
       imageDelegate: widget.imageDelegate,
       physics: widget.physics,
+      keyboardAppearance: widget.keyboardAppearance,
     );
 
     if (widget.height != null) {
@@ -75,12 +85,11 @@ class _ZefyrFieldState extends State<ZefyrField> {
   }
 
   InputDecoration _getEffectiveDecoration() {
-    final InputDecoration effectiveDecoration =
-        (widget.decoration ?? const InputDecoration())
-            .applyDefaults(Theme.of(context).inputDecorationTheme)
-            .copyWith(
-              enabled: widget.mode == ZefyrMode.edit,
-            );
+    final effectiveDecoration = (widget.decoration ?? const InputDecoration())
+        .applyDefaults(Theme.of(context).inputDecorationTheme)
+        .copyWith(
+          enabled: _effectiveMode == ZefyrMode.edit,
+        );
 
     return effectiveDecoration;
   }
