@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
-import 'dart:math' as math;
 import 'dart:ui' as ui;
 
 import 'package:flutter/rendering.dart';
@@ -11,6 +10,8 @@ import 'package:meta/meta.dart';
 import 'package:notus/notus.dart';
 
 import 'editable_box.dart';
+import 'selection_utils.dart';
+import 'theme.dart';
 
 /// Provides interface for embedding images into Zefyr editor.
 // TODO: allow configuring image sources and related toolbar buttons.
@@ -58,9 +59,13 @@ class _ZefyrImageState extends State<ZefyrImage> {
 
   @override
   Widget build(BuildContext context) {
+    final theme = ZefyrTheme.of(context);
     final image = widget.delegate.buildImage(context, imageSource);
     return _EditableImage(
-      child: image,
+      child: Padding(
+        padding: theme.defaultLineTheme.padding,
+        child: image,
+      ),
       node: widget.node,
     );
   }
@@ -88,25 +93,19 @@ class _EditableImage extends SingleChildRenderObjectWidget {
 class RenderEditableImage extends RenderBox
     with RenderObjectWithChildMixin<RenderBox>, RenderProxyBoxMixin<RenderBox>
     implements RenderEditableBox {
-  static const kPaddingBottom = 24.0;
-
   RenderEditableImage({
     RenderImage child,
     @required EmbedNode node,
-  }) : _node = node {
+  }) : node = node {
     this.child = child;
   }
 
   @override
-  EmbedNode get node => _node;
-  EmbedNode _node;
-  set node(EmbedNode value) {
-    _node = value;
-  }
+  EmbedNode node;
 
   // TODO: Customize caret height offset instead of adjusting here by 2px.
   @override
-  double get preferredLineHeight => size.height - kPaddingBottom + 2.0;
+  double get preferredLineHeight => size.height + 2.0;
 
   @override
   SelectionOrder get selectionOrder => SelectionOrder.foreground;
@@ -115,22 +114,18 @@ class RenderEditableImage extends RenderBox
   TextSelection getLocalSelection(TextSelection documentSelection) {
     if (!intersectsWithSelection(documentSelection)) return null;
 
-    int nodeBase = node.documentOffset;
-    int nodeExtent = nodeBase + node.length;
-    int base = math.max(0, documentSelection.baseOffset - nodeBase);
-    int extent =
-        math.min(documentSelection.extentOffset, nodeExtent) - nodeBase;
-    return documentSelection.copyWith(baseOffset: base, extentOffset: extent);
+    final nodeBase = node.documentOffset;
+    final nodeExtent = nodeBase + node.length;
+    return selectionRestrict(nodeBase, nodeExtent, documentSelection);
   }
 
   @override
   List<ui.TextBox> getEndpointsForSelection(TextSelection selection) {
-    TextSelection local = getLocalSelection(selection);
+    final local = getLocalSelection(selection);
     if (local.isCollapsed) {
       final dx = local.extentOffset == 0 ? _childOffset.dx : size.width;
       return [
-        ui.TextBox.fromLTRBD(
-            dx, 0.0, dx, size.height - kPaddingBottom, TextDirection.ltr),
+        ui.TextBox.fromLTRBD(dx, 0.0, dx, size.height, TextDirection.ltr),
       ];
     }
 
@@ -145,7 +140,7 @@ class RenderEditableImage extends RenderBox
 
   @override
   TextPosition getPositionForOffset(Offset offset) {
-    int position = _node.documentOffset;
+    var position = node.documentOffset;
 
     if (offset.dx > size.width / 2) {
       position++;
@@ -155,21 +150,21 @@ class RenderEditableImage extends RenderBox
 
   @override
   TextRange getWordBoundary(TextPosition position) {
-    final start = _node.documentOffset;
+    final start = node.documentOffset;
     return TextRange(start: start, end: start + 1);
   }
 
   @override
   bool intersectsWithSelection(TextSelection selection) {
-    final int base = node.documentOffset;
-    final int extent = base + node.length;
-    return base <= selection.extentOffset && selection.baseOffset <= extent;
+    final base = node.documentOffset;
+    final extent = base + node.length;
+    return selectionIntersectsWith(base, extent, selection);
   }
 
   @override
   Offset getOffsetForCaret(TextPosition position, Rect caretPrototype) {
     final pos = position.offset - node.documentOffset;
-    Offset caretOffset = _childOffset - Offset(kHorizontalPadding, 0.0);
+    var caretOffset = _childOffset - Offset(kHorizontalPadding, 0.0);
     if (pos == 1) {
       caretOffset =
           caretOffset + Offset(_lastChildSize.width + kHorizontalPadding, 0.0);
@@ -183,7 +178,7 @@ class RenderEditableImage extends RenderBox
     final localSelection = getLocalSelection(selection);
     assert(localSelection != null);
     if (!localSelection.isCollapsed) {
-      final Paint paint = Paint()
+      final paint = Paint()
         ..color = selectionColor
         ..style = PaintingStyle.stroke
         ..strokeWidth = 3.0;
@@ -193,6 +188,7 @@ class RenderEditableImage extends RenderBox
     }
   }
 
+  @override
   void paint(PaintingContext context, Offset offset) {
     super.paint(context, offset + _childOffset);
   }
@@ -203,7 +199,7 @@ class RenderEditableImage extends RenderBox
 
   Offset get _childOffset {
     final dx = (size.width - _lastChildSize.width) / 2 + kHorizontalPadding;
-    final dy = (size.height - _lastChildSize.height - kPaddingBottom) / 2;
+    final dy = (size.height - _lastChildSize.height) / 2;
     return Offset(dx, dy);
   }
 
@@ -226,7 +222,7 @@ class RenderEditableImage extends RenderBox
       );
       child.layout(childConstraints, parentUsesSize: true);
       _lastChildSize = child.size;
-      size = Size(constraints.maxWidth, _lastChildSize.height + kPaddingBottom);
+      size = Size(constraints.maxWidth, _lastChildSize.height);
     } else {
       performResize();
     }
