@@ -24,15 +24,16 @@ class InputConnectionController implements TextInputClient {
 
   /// Opens or closes input connection based on the current state of
   /// [focusNode] and [value].
-  void openOrCloseConnection(FocusNode focusNode, TextEditingValue value) {
+  void openOrCloseConnection(FocusNode focusNode, TextEditingValue value,
+      Brightness keyboardAppearance) {
     if (focusNode.hasFocus && focusNode.consumeKeyboardToken()) {
-      openConnection(value);
+      openConnection(value, keyboardAppearance);
     } else if (!focusNode.hasFocus) {
       closeConnection();
     }
   }
 
-  void openConnection(TextEditingValue value) {
+  void openConnection(TextEditingValue value, Brightness keyboardAppearance) {
     if (!hasConnection) {
       _lastKnownRemoteTextEditingValue = value;
       _textInputConnection = TextInput.attach(
@@ -42,6 +43,7 @@ class InputConnectionController implements TextInputClient {
           obscureText: false,
           autocorrect: true,
           inputAction: TextInputAction.newline,
+          keyboardAppearance: keyboardAppearance,
           textCapitalization: TextCapitalization.sentences,
         ),
       )..setEditingState(value);
@@ -79,7 +81,7 @@ class InputConnectionController implements TextInputClient {
 
     if (actualValue == _lastKnownRemoteTextEditingValue) return;
 
-    bool shouldRemember = value.text != _lastKnownRemoteTextEditingValue.text;
+    final shouldRemember = value.text != _lastKnownRemoteTextEditingValue.text;
     _lastKnownRemoteTextEditingValue = actualValue;
     _textInputConnection.setEditingState(actualValue);
     if (shouldRemember) {
@@ -156,6 +158,25 @@ class InputConnectionController implements TextInputClient {
     }
   }
 
+  @override
+  TextEditingValue get currentTextEditingValue =>
+      _lastKnownRemoteTextEditingValue;
+
+  @override
+  void updateFloatingCursor(RawFloatingCursorPoint point) {
+    // TODO: implement updateFloatingCursor
+  }
+
+  @override
+  void connectionClosed() {
+    if (hasConnection) {
+      _textInputConnection.connectionClosedReceived();
+      _textInputConnection = null;
+      _lastKnownRemoteTextEditingValue = null;
+      _sentRemoteValues.clear();
+    }
+  }
+
   //
   // Private members
   //
@@ -164,8 +185,12 @@ class InputConnectionController implements TextInputClient {
   TextInputConnection _textInputConnection;
   TextEditingValue _lastKnownRemoteTextEditingValue;
 
+  // TODO: figure out if we need to support autofill
   @override
-  void updateFloatingCursor(RawFloatingCursorPoint point) {
-    // TODO: implement updateFloatingCursor
+  AutofillScope get currentAutofillScope => null;
+
+  @override
+  void showAutocorrectionPromptRect(int start, int end) {
+    // TODO: implement showAutocorrectionPromptRect
   }
 }
