@@ -11,9 +11,9 @@ void main() {
       var newText = 'fasterDiff';
       var result = fastDiff(oldText, newText, 6);
       expect(result.start, 4);
-      expect(result.deleted, "");
-      expect(result.inserted, "er");
-      expect("$result", 'DiffResult[4, "", "er"]');
+      expect(result.deleted, '');
+      expect(result.inserted, 'er');
+      expect('$result', 'DiffResult[4, "", "er"]');
     });
 
     test('delete', () {
@@ -21,8 +21,8 @@ void main() {
       var newText = 'fasDiff';
       var result = fastDiff(oldText, newText, 3);
       expect(result.start, 3);
-      expect(result.deleted, "t");
-      expect(result.inserted, "");
+      expect(result.deleted, 't');
+      expect(result.inserted, '');
     });
 
     test('replace', () {
@@ -30,8 +30,8 @@ void main() {
       var newText = 'fas_Diff';
       var result = fastDiff(oldText, newText, 4);
       expect(result.start, 3);
-      expect(result.deleted, "t");
-      expect(result.inserted, "_");
+      expect(result.deleted, 't');
+      expect(result.inserted, '_');
     });
   });
 }
